+++ spec/acceptance/amazon_spec.rb
@@ -294,7 +294,7 @@ RSpec.describe 'Amazon' do
       context = WSDL::Request::DSLContext.new(
         document:,
         security: WSDL::Security::Config.new,
-        limits: WSDL.limits
+        limits: WSDL::Limits.new
       )
 
       context.instance_exec do